seed-story/requirements.txt
# SEED-Story Docker Requirements
# Based on original requirements.txt with additional optimizations

# Core ML/AI packages
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
accelerate==0.25.0
transformers==4.35.2
diffusers==0.21.4
xformers==0.0.23.post1
triton==2.1.0

# Computer Vision
opencv-python-headless==4.9.0.80
Pillow==10.2.0
imageio==2.34.0
scikit-image==0.21.0
easyocr==1.7.1

# ML utilities
numpy==1.23.5
scipy==1.10.1
matplotlib==3.7.5
seaborn==0.13.2
pandas==2.0.3
PyWavelets==1.4.1

# Model training and optimization
pytorch-lightning==2.2.1
deepspeed==0.9.5
peft==0.4.0
safetensors==0.4.2

# Hugging Face ecosystem
# Let pip resolve huggingface-hub version automatically
tokenizers>=0.15.0
sentencepiece==0.2.0

# Configuration and utilities
hydra-core==1.3.2
omegaconf==2.3.0
pyrootutils==1.0.4
PyYAML==6.0.1
click==8.1.7
tqdm==4.66.2

# Web and API
Flask==3.0.2
requests==2.31.0
Werkzeug==3.0.1

# File handling
braceexpand==0.1.7
filelock==3.9.0
packaging==24.0
protobuf==4.21.2
hjson==3.1.0

# Monitoring and logging
wandb==0.16.4
tensorboard==2.14.0
sentry-sdk==1.43.0

# Additional utilities
regex==2023.12.25
Markdown==3.6
thop==0.1.1.post2209072238
ninja==1.11.1.1
shapely==2.0.3
setproctitle==1.3.3
torchdata==0.7.1
